// FCAPROGRAMACION/FrontEnd/MIGRACION_CPLCAP005/src/app/models/common/cplcap003.ts
export class ParametrosCPLCAP003 {
  resistenciaAfines: boolean;
  refileMaximo: number;
  refileMinimo: number;
  diasAdelantadoFE: number;
  todosAnchos: boolean;
  anchosCalculo: number;
  largoMinProgr: number;
  aumentoPedido: number;
  scoresMaximos: number;
}

export class DatosComboCPLCAP003 {
  codigo: number;
  descripcion: string;
  cantidad: number;
}